change pml_size to a signed type to flag the error checking already present but which would never work previously from delving into negative values
git-svn-id: svn+ssh://frodo/cluster/svn/projects@5313 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/include/psc_util/mlist.h
@@ -34,7 +34,7 @@ struct psc_mlist {
 	struct psclist_head	pml_index_lentry;	/* link adjoining mlists */
 	psc_spinlock_t		pml_lock;
 
-	size_t			pml_size;
+	int			pml_size;
 	struct psclist_head	pml_listhd;
 	struct multilock_cond	pml_mlcond_empty;	/* wait here while list is empty */
 	char			pml_name[PML_NAME_MAX];	/* for ml mgmt */
